Reformatted and implemented jdbc for role. Ran all tests

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -3,7 +3,10 @@ package com.ust.pos.user;
 import com.ust.pos.dto.UserDto;
 import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
+import org.jspecify.annotations.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -11,6 +14,11 @@ import org.springframework.web.bind.annotation.*;
 @Controller
 @RequestMapping("/user")
 public class UserController {
+
+    private static final String USER_LIST = "user/list";
+    private static final String USER_VIEW = "user/user";
+    private static final String REDIRECT_USER_LIST = "redirect:/user/list";
+    private static final String REDIRECT_USER_LOGIN = "redirect:/login";
 
     @Autowired
     private UserService userService;
@@ -21,7 +29,7 @@ public class UserController {
     @GetMapping("/list")
     public String home(Model model) {
         model.addAttribute("users", userService.findAll(null));
-        return "user/list";
+        return USER_LIST;
     }
 
     @GetMapping("/get")
@@ -29,7 +37,7 @@ public class UserController {
         UserDto response = userService.findByUserName(username);
         model.addAttribute("userDto", response);
         model.addAttribute("roles", roleService.findAll(null));
-        return "user/user";
+        return USER_VIEW;
     }
 
     @PostMapping("/update")
@@ -38,12 +46,20 @@ public class UserController {
         if (!response.isSuccess()) {
             model.addAttribute("message", response.getMessage());
         }
-        return "redirect:/user/list";
+        return REDIRECT_USER_LIST;
     }
 
     @GetMapping("/delete")
     public String delete(@RequestParam String username) {
-        userService.delete(username);
-        return "redirect:/user/list";
+        @Nullable Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null) {
+            String loggedInUser = authentication.getName();
+            userService.delete(username);
+            if (loggedInUser.equals(username)) {
+                SecurityContextHolder.clearContext();
+                return REDIRECT_USER_LOGIN;
+            }
+        }
+        return REDIRECT_USER_LIST;
     }
 }
